+++ payments/RELEASES.md
@@ -68,3 +68,13 @@ Based on:
 - [typescript v2.0.2] payments
 ### Releases
 - [NPM v2.0.2] https://www.npmjs.com/package/@wingspan/payments/v/2.0.2 - payments
+
+## 2024-02-23 20:38:04
+### Changes
+Based on:
+- OpenAPI Doc 1.0.0-oas3.1 ./openapi/payment_speakeasy_edit.yaml
+- Speakeasy CLI 1.193.1 (2.272.1) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v2.1.0] payments
+### Releases
+- [NPM v2.1.0] https://www.npmjs.com/package/@wingspan/payments/v/2.1.0 - payments